+++ domain/src/test/java/com/thomaskioko/paybillmanager/domain/interactor/bills/GetBillsTest.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.paybillmanager.domain.interactor
+package com.thomaskioko.paybillmanager.domain.interactor.bills
 
 import com.nhaarman.mockitokotlin2.whenever
 import com.thomaskioko.paybillmanager.domain.bills.GetBills
